+++ packages/components/tabs/__tests__/tabs.test.tsx
@@ -307,7 +307,6 @@ describe('Tabs 组件测试', () => {
               value: 0,
             },
           ]}
-          tabClick={() => ''}
         ></TabNav>
       </div>,
     );
@@ -317,7 +316,7 @@ describe('Tabs 组件测试', () => {
   });
 
   const mockFn = vi.spyOn(HTMLDivElement.prototype, 'getBoundingClientRect');
-  mockFn.mockImplementation(() => ({ width: 20, x: 5, clientX: 5 }));
+  mockFn.mockImplementation(() => ({ width: 20, x: 5, clientX: 5 }) as any);
 
   test('test drag', async () => {
     const onDragSort = vi.fn(() => {
@@ -357,8 +356,8 @@ describe('Tabs 组件测试', () => {
         targetIndex: 0,
       },
     });
-    expect(onDragSort).toHaveBeenCalled(1);
-    expect(onDragSort.mock.calls[0][0].target.value).toEqual('vue');
+    expect(onDragSort).toHaveBeenCalled();
+    expect((onDragSort as any).mock.calls[0][0].target.value).toEqual('vue');
     expect(container.querySelectorAll('.t-tabs__nav-item-text-wrapper').item(0).firstChild.nodeValue).toEqual('react');
   });