+++ test/utils/index.jsx
@@ -1,7 +1,8 @@
 import '@testing-library/jest-dom';
+
+import { vi } from 'vitest';
 import { act, createEvent, fireEvent } from '@testing-library/react';
 import _userEvent from '@testing-library/user-event';
-import { vi } from 'vitest';
 
 import { EVENTS_MAP } from './events';
 
@@ -175,10 +176,11 @@ export function mockIntersectionObserver(mockData, mockFunc) {
     },
   ];
 
-  window.IntersectionObserver = vi.fn((callback, { root, rootMargin }) => ({
+  window.IntersectionObserver = vi.fn((callback, { root, rootMargin, scrollMargin }) => ({
     root,
     rootMargin,
     thresholds,
+    scrollMargin,
     observe: observe
       ? (element) => {
           observe(element, callback);