+++ packages/tdesign-react/CHANGELOG.md
@@ -5,6 +5,54 @@ toc: false
 spline: explain
 ---
 
+## 🌈 1.18.2 `2026-08-14`
+
+### 🚀 Features
+
+- `Comment`: `actions` 支持单个 `TNode` @RSS1102 ([#4354](https://github.com/Tencent/tdesign-react/pull/4354))
+- `Form`: 末尾表单项增加 `--last` 标识，保证 `margin` 样式正常生效 @uyarn ([#4356](https://github.com/Tencent/tdesign-react/pull/4356))
+- `Tag`: `title` 支持 `Boolean` 类型，传入 `false` 则不会显示 `title`，不传或传入 `true` 则显示默认的内容。 @Wesley-0808 ([#4336](https://github.com/Tencent/tdesign-react/pull/4336))
+
+### 🐞 Bug Fixes
+
+- `Button`: 
+    - 移除组件不合理的默认 `margin:0` 和  `z-index:0` @RylanBot ([common#2636](https://github.com/Tencent/tdesign-common/pull/2636))
+   - 修复自定义 SVG 图标与文字之间缺少间距的问题 @xy200303 ([common#2639](https://github.com/Tencent/tdesign-common/pull/2639)) 
+- `Comment`: 修复操作项图标与文字未垂直居中的问题 @RSS1102 ([#4354](https://github.com/Tencent/tdesign-react/pull/4354))
+- `ConfigProvider`: 修复 `table.sortIcon` 不生效的问题 @RylanBot ([#4338](https://github.com/Tencent/tdesign-react/pull/4338))
+- `EnhancedTable`: 
+  - 修复半选状态异常的问题 @RylanBot ([#4104](https://github.com/Tencent/tdesign-react/pull/4104))
+  - 修复非受控且 `tree.checkStrictly` 为 `false` 时，选中父节点后子节点没有同步更新的问题 @RylanBot ([#4104](https://github.com/Tencent/tdesign-react/pull/4104))
+- `Form`: 修复嵌套表单在触发 `validate` 后重渲染导致数值丢失的问题 @RylanBot ([#4350](https://github.com/Tencent/tdesign-react/pull/4350))
+- `Popup`: 修复鼠标在浮层上左键长按或右键点击后，移出浮层无法自动关闭的问题 @RylanBot ([#4287](https://github.com/Tencent/tdesign-react/pull/4287))
+- `SelectInput`: 修复 `1.18.1` 的调整，导致单选且未开启 `filterable` 时，`valueDisplay` 为自定义字符串导致的 DOM 结构变更问题 @RylanBot ([#4351](https://github.com/Tencent/tdesign-react/pull/4351))
+- `Steps`: 
+  - 修复 `layout='vertical'` 时，连接线不对齐和上下图标距离不一致的问题 @RylanBot ([common#2670](https://github.com/Tencent/tdesign-common/pull/2670))
+  - 修复默认和选中态的连接线粗细不一致的问题 @RylanBot ([common#2670](https://github.com/Tencent/tdesign-common/pull/2670))
+- `Table`: 
+    - 修复全选会篡改禁用项选中状态的问题 @RylanBot ([#4104](https://github.com/Tencent/tdesign-react/pull/4104))
+    - 修复过滤后的数据量在跨越虚拟滚动的 threshold 时，筛选器下拉框被强行关闭的问题 @RylanBot ([#3803](https://github.com/Tencent/tdesign-react/pull/3803))
+
+## 🌈 1.18.1 `2026-07-17`
+
+### 🚀 Features
+
+- `SelectInput`: 支持 `focus` 时的内容为 `valueDisplay` 的对应值。 @RylanBot ([#3891](https://github.com/Tencent/tdesign-react/pull/3891))
+
+### 🐞 Bug Fixes
+
+- `Button`: 优化斜八度的动画效果 @uyarn ([#4331](https://github.com/Tencent/tdesign-react/pull/4331))
+- `ColorPicker`: 修复渐变角度的 `title` 格式异常的问题 @RylanBot ([#4324](https://github.com/Tencent/tdesign-react/pull/4324))
+- `Menu`: 修复弹出窗口位置在左侧宽度不够向右侧翻转时的间距错误问题 @RSS1102 ([common#2631](https://github.com/Tencent/tdesign-common/pull/2631))
+- `Select`: 
+  - 修复 `label` 为非 `string` 类型时，开启 `filterable` 搜索后报错的问题。 @RylanBot ([#3891](https://github.com/Tencent/tdesign-react/pull/3891))
+  - 修复使用 `label` / `content` / `children` 自定义元素时，下拉框选项的高度没有撑开的问题。  ([common#2537](https://github.com/Tencent/tdesign-common/pull/2537)) @RylanBot
+- `SelectInput`: 
+  - 修复单选且 `valueDisplay` 为自定义元素时，开启 `filterable` 无法显示输入内容的问题。 @RylanBot ([#3891](https://github.com/Tencent/tdesign-react/pull/3891))
+  - 修复 `inputProps.inputClass` 无法生效的问题。先前 `inputProps.className` 被错误地应用到了 `t-class` 同级位置，导致正确的 `inputProps.inputClass` 配置无效，之前使用相关属性的业务注意此变更 ⚠️ @RylanBot ([#3891](https://github.com/Tencent/tdesign-react/pull/3891))
+- `Table`: 在非 Chromium 内核浏览器时，由于 `min-width` 兼容性问题，降级到 `width` @RylanBot ([#4219](https://github.com/Tencent/tdesign-react/pull/4219))
+- `TagInput`: 修复带 label 且已有 tag 时输入内容垂直对齐偏差的问题 @DaZuiZui  ([common#2564](https://github.com/Tencent/tdesign-common/pull/2564))
+
 ## 🌈 1.18.0 `2026-06-26`
 
 ### 🚀 Features